% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_kevin}
\alias{estimate_kevin}
\title{Estimate the largest directional difference in means}
\usage{
estimate_kevin(y)
}
\arguments{
\item{y}{data vector}
}
\value{
an index (positive integer)
}
\description{
Estimate the largest directional difference in means
}
