% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poly.pval_kevin}
\alias{poly.pval_kevin}
\title{Compute pvalue from polyhedron}
\usage{
poly.pval_kevin(mat, y, sigma2, contrast)
}
\arguments{
\item{mat}{Gamma matrix}

\item{y}{data vector}

\item{sigma2}{variance of the generative process of y}

\item{contrast}{vector}
}
\value{
a list of the pvalue, the numerator of the pvalue, and the denominator of the pvalue
}
\description{
Compute pvalue from polyhedron
}
