% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{polyhedron_kevin}
\alias{polyhedron_kevin}
\title{Compute polyhedraon of \code{estimate_kevin}}
\usage{
polyhedron_kevin(y, i)
}
\arguments{
\item{y}{data vector}

\item{i}{index where \code{estimate_kevin} estimated}
}
\value{
a matrix to be the Gamma matrix
}
\description{
Compute polyhedraon of \code{estimate_kevin}
}
