% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampler_kevin}
\alias{sampler_kevin}
\title{Sampler for randomized inference}
\usage{
sampler_kevin(y, sigma02, sigma2, num_trials, contrast)
}
\arguments{
\item{y}{data vector}

\item{sigma02}{variance of the randomized noise being added}

\item{sigma2}{variance of the generative process of y}

\item{num_trials}{number of trials for the sampler}

\item{contrast}{vector}
}
\value{
pvalue
}
\description{
Sampler for randomized inference
}
